陈罗门[实现了资费管理页面的修改操作]

+++ FeeApp/models.py
@@ -5,7 +5,7 @@ from django.db import models
 class Cost(models.Model):
     name = models.CharField(max_length=32)
     base_duration = models.CharField(max_length=32, null=True)
-    base_cost = models.IntegerField(null=True)
+    base_cost = models.FloatField(null=True)
     unit_cost = models.FloatField(null=True)
     status = models.NullBooleanField(default=False)
     descr = models.CharField(max_length=256)
